+++ src/main/java/co/ipregistry/api/client/model/Carrier.java
@@ -17,10 +17,12 @@
 package co.ipregistry.api.client.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
+@EqualsAndHashCode
 @Getter
 @NoArgsConstructor
 @ToString