+++ src/main/java/co/ipregistry/api/client/model/Continent.java
@@ -21,6 +21,7 @@ import lombok.*;
 
 @AllArgsConstructor
 @Builder
+@EqualsAndHashCode
 @Getter
 @NoArgsConstructor
 @ToString