+++ src/main/java/co/ipregistry/api/client/model/RequesterIpInfo.java
@@ -18,10 +18,12 @@ package co.ipregistry.api.client.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 @AllArgsConstructor
+@EqualsAndHashCode
 @NoArgsConstructor
 @ToString(callSuper = true)
 public class RequesterIpInfo extends IpInfo {