+++ src/main/java/co/ipregistry/api/client/model/UserAgentList.java
@@ -17,12 +17,10 @@
 package co.ipregistry.api.client.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 @AllArgsConstructor
+@EqualsAndHashCode
 @NoArgsConstructor
 @Getter
 @Setter