+++ src/main/java/co/ipregistry/api/client/model/UserAgentOperatingSystem.java
@@ -17,12 +17,10 @@
 package co.ipregistry.api.client.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @AllArgsConstructor
+@EqualsAndHashCode
 @Builder
 @Data
 @NoArgsConstructor