+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Implement equals/hashCode in all IpInfo subfield classes.
 - Default cache implementation will expire entries older than 24 hours by default.
+- Allow passing any int value to `expireAfter` and `maximumSize` parameters of DefaultCache constructor.
 
 ## [2.0.0] - 2019-07-19
 ### Added